Resolved conflicts.
